ENH: Adding an API for interpolation functions and derivatives that makes it possible to implement for cell like vtkPolygon. For all other cell makeing the virtual function call the static one (for now).

+++ Common/vtkWin32ProcessOutputWindow.h
@@ -40,7 +40,7 @@ protected:
   ~vtkWin32ProcessOutputWindow();
 
   int Initialize();
-  void Write(const char* data, int length);
+  void Write(const char* data, size_t length);
 
   // The write end of the pipe to the child process.
   vtkWindowsHANDLE OutputPipe;
